requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
deprecation==2.1.0
distro==1.9.0
frozenlist==1.5.0
gotrue==2.12.0
greenlet==3.2.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.53
langchain-openai==0.3.13
langchain-text-splitters==0.3.8
langsmith==0.3.31
marshmallow==3.26.1
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.75.0
orjson==3.10.16
packaging==24.2
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
PyJWT==2.10.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.0
PyYAML==6.0.2
realtime==2.4.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==14.2
yarl==1.20.0
zstandard==0.23.0
